Correccion de Codigos 4
Se corrigieron los Códigos, modificando la globalidad de la tabla de símbolos, logrando hacer que imprima null
