+++ .vscode/launch.json
@@ -8,31 +8,27 @@
             "type": "java",
             "name": "Principal",
             "request": "launch",
-            "mainClass": "Principal",
+            "mainClass": "AnalizadorSintacticoFinal.Principal",
             "projectName": "Compiladores5CV2Elbueno_f7641fb8",
-          "args": ["C:\\Users\\ezequ\\OneDrive\\Desktop\\Compiladores5CV2Elbueno\\AnalizadorSintacticoFinal\\Prueba.txt"]
-          //"args": ["C:\\Users\\ezequ\\OneDrive\\Desktop\\Compiladores5CV2Elbueno\\AnalizadorSintacticoFinal\\Prueba2.txt"]
-          //"args": ["C:\\Users\\ezequ\\OneDrive\\Desktop\\Compiladores5CV2Elbueno\\AnalizadorSintacticoFinal\\Prueba3.txt"]
-          //"args": ["C:\\Users\\ezequ\\OneDrive\\Desktop\\Compiladores5CV2Elbueno\\AnalizadorSintacticoFinal\\Prueba4.txt"]
-          //"args": ["C:\\Users\\ezequ\\OneDrive\\Desktop\\Compiladores5CV2Elbueno\\AnalizadorSintacticoFinal\\Prueba5.txt"]
-          //"args": ["C:\\Users\\ezequ\\OneDrive\\Desktop\\Compiladores5CV2Elbueno\\AnalizadorSintacticoFinal\\Prueba6.txt"]
-          //"args": ["C:\\Users\\ezequ\\OneDrive\\Desktop\\Compiladores5CV2Elbueno\\AnalizadorSintacticoFinal\\Prueba7.txt"]
-          //"args": ["C:\\Users\\ezequ\\OneDrive\\Desktop\\Compiladores5CV2Elbueno\\AnalizadorSintacticoFinal\\Prueba8.txt"]
-          //"args": ["C:\\Users\\ezequ\\OneDrive\\Desktop\\Compiladores5CV2Elbueno\\AnalizadorSintacticoFinal\\Prueba9.txt"]
-          //"args": ["C:\\Users\\ezequ\\OneDrive\\Desktop\\Compiladores5CV2Elbueno\\AnalizadorSintacticoFinal\\Prueba10.txt"]
+            //"args": ["C:\\Users\\ezequ\\OneDrive\\Desktop\\Compiladores5CV2Elbueno\\AnalizadorSintacticoFinal\\AnalizadorSintacticoFinal\\Prueba.txt"]
+            //"args": ["C:\\Users\\ezequ\\OneDrive\\Desktop\\Compiladores5CV2Elbueno\\AnalizadorSintacticoFinal\\AnalizadorSintacticoFinal\\Prueba2.txt"]
+            //"args": ["C:\\Users\\ezequ\\OneDrive\\Desktop\\Compiladores5CV2Elbueno\\AnalizadorSintacticoFinal\\AnalizadorSintacticoFinal\\Prueba3.txt"]
+            //"args": ["C:\\Users\\ezequ\\OneDrive\\Desktop\\Compiladores5CV2Elbueno\\AnalizadorSintacticoFinal\\AnalizadorSintacticoFinal\\Prueba4.txt"]
+            //"args": ["C:\\Users\\ezequ\\OneDrive\\Desktop\\Compiladores5CV2Elbueno\\AnalizadorSintacticoFinal\\AnalizadorSintacticoFinal\\Prueba5.txt"]
+            //"args": ["C:\\Users\\ezequ\\OneDrive\\Desktop\\Compiladores5CV2Elbueno\\AnalizadorSintacticoFinal\\AnalizadorSintacticoFinal\\Prueba6.txt"]
+            //"args": ["C:\\Users\\ezequ\\OneDrive\\Desktop\\Compiladores5CV2Elbueno\\AnalizadorSintacticoFinal\\AnalizadorSintacticoFinal\\Prueba7.txt"]
+            "args": ["C:\\Users\\ezequ\\OneDrive\\Desktop\\Compiladores5CV2Elbueno\\AnalizadorSintacticoFinal\\AnalizadorSintacticoFinal\\Prueba8.txt"]
+            //"args": ["C:\\Users\\ezequ\\OneDrive\\Desktop\\Compiladores5CV2Elbueno\\AnalizadorSintacticoFinal\\AnalizadorSintacticoFinal\\Prueba9.txt"]
+            //"args": ["C:\\Users\\ezequ\\OneDrive\\Desktop\\Compiladores5CV2Elbueno\\AnalizadorSintacticoFinal\\AnalizadorSintacticoFinal\\Prueba10.txt"]
+            
         },
+        
         {
             "type": "java",
             "name": "Current File",
             "request": "launch",
             "mainClass": "${file}"
         },
-        {
-            "type": "java",
-            "name": "Principal",
-            "request": "launch",
-            "mainClass": "mx.ipn.escom.compiladores.Principal",
-            "projectName": "Compiladores5CV2Elbueno_f7641fb8"
-        }
+        
     ]
 }
