+++ AnalizadorSintacticoFinal/AnalizadorSintacticoFinal/Nodo.java
@@ -1,3 +1,5 @@
+package AnalizadorSintacticoFinal;
+
 import java.util.ArrayList;
 import java.util.List;
 
@@ -5,45 +7,43 @@ public class Nodo {
     private final Token value;
     private List<Nodo> hijos;
 
-    public Nodo(Token value){
+    public Nodo(Token value) {
         this.value = value;
     }
 
-    public void insertarHijo(Nodo n){
-        if(hijos == null){
+    public void insertarHijo(Nodo n) {
+        if (hijos == null) {
             hijos = new ArrayList<>();
             hijos.add(n);
-        }
-        else{
+        } else {
             hijos.add(0, n);
         }
     }
 
-    public void insertarSiguienteHijo(Nodo n){
-        if(hijos == null){
+    public void insertarSiguienteHijo(Nodo n) {
+        if (hijos == null) {
             hijos = new ArrayList<>();
             hijos.add(n);
-        }
-        else{
+        } else {
             hijos.add(n);
         }
     }
 
-    public void insertarHijos(List<Nodo> nodosHijos){
-        if(hijos == null){
+    public void insertarHijos(List<Nodo> nodosHijos) {
+        if (hijos == null) {
             hijos = new ArrayList<>();
         }
 
-        for(Nodo n : nodosHijos){
+        for (Nodo n : nodosHijos) {
             hijos.add(n);
         }
     }
 
-    public Token getValue(){
+    public Token getValue() {
         return value;
     }
 
-    public List<Nodo> getHijos(){
+    public List<Nodo> getHijos() {
         return hijos;
     }
 }
