+++ AnalizadorSintacticoFinal/AnalizadorSintacticoFinal/Prueba.txt
@@ -1,0 +1,5 @@
+// Prueba 1:
+var nombre = "Gabriel";
+var apellido = "Rodriguez";
+
+print nombre + " " + apellido;
