+++ AnalizadorSintacticoFinal/AnalizadorSintacticoFinal/Prueba10.txt
