+++ AnalizadorSintacticoFinal/AnalizadorSintacticoFinal/SolverAritmetico.java
@@ -1,56 +1,55 @@
-
-
+package AnalizadorSintacticoFinal;
 
 public class SolverAritmetico {
 
-private Nodo nodo;
+    private final Nodo nodo;
 
     public SolverAritmetico(Nodo nodo) {
         this.nodo = nodo;
     }
-    public SolverAritmetico(){}
 
     public Object resolver(){
         return resolver(nodo);
     }
-    public Object resolver(Nodo n){
+    private Object resolver(Nodo n)
+    {
         // No tiene hijos, es un operando
-        if(n.getHijos() == null){
-            if(n.getValue().tipo == TipoToken.NUMERO || n.getValue().tipo == TipoToken.CADENA) {
+        if(n.getHijos() == null)
+        {
+            if (n.getValue().tipo == TipoToken.NUMERO || n.getValue().tipo == TipoToken.CADENA)
+            {
                 return n.getValue().literal;
             }
-            else if(n.getValue().tipo == TipoToken.IDENTIFICADOR){
+            else if (n.getValue().tipo == TipoToken.IDENTIFICADOR)
+            {
                 // Ver la tabla de símbolos
-                if(TablaDeSimbolos.existeIdentificador(n.getValue().lexema)){
+                if (TablaDeSimbolos.existeIdentificador(n.getValue().lexema) == true)
+                {
                     return TablaDeSimbolos.obtener(n.getValue().lexema);
-                }else {
-                    System.err.println("Error : La variable " +  n.getValue().lexema + " no esta definida ");
+                }
+                else
+                {
+                    System.out.println("La variable " +  n.getValue().lexema + " no existe");
                     System.exit(1);
                 }
-            } else if (n.getValue().tipo == TipoToken.FALSE) {
+            }
+            else if (n.getValue().tipo == TipoToken.FALSE)
+            {
                 return false;
-            } else if (n.getValue().tipo == TipoToken.TRUE) {
+            }
+            else if (n.getValue().tipo == TipoToken.TRUE)
+            {
                 return true;
             }
         }
-
         // Por simplicidad se asume que la lista de hijos del nodo tiene dos elementos
         Nodo izq = n.getHijos().get(0);
         Nodo der = n.getHijos().get(1);
 
-
         Object resultadoIzquierdo = resolver(izq);
         Object resultadoDerecho = resolver(der);
 
-
-        if (resultadoIzquierdo instanceof Boolean && resultadoDerecho instanceof Boolean){
-            switch (n.getValue().tipo){
-                case AND:
-                    return ((Boolean)resultadoIzquierdo && (Boolean) resultadoDerecho);
-                case OR:
-                    return ((Boolean)resultadoIzquierdo || (Boolean) resultadoDerecho);
-            }
-        }else if(resultadoIzquierdo instanceof Double && resultadoDerecho instanceof Double){
+        if(resultadoIzquierdo instanceof Double && resultadoDerecho instanceof Double){
             switch (n.getValue().tipo){
                 case SUMA:
                     return ((Double)resultadoIzquierdo + (Double) resultadoDerecho);
@@ -73,38 +72,49 @@ private Nodo nodo;
                 case DIFERENTE:
                     return (!((Double) resultadoIzquierdo).equals((Double) resultadoDerecho));
                 case IGUAL:
-                    // Asignar el valor de la derecha a la variable de la izquierda
-                    if (izq.getValue().tipo == TipoToken.IDENTIFICADOR){
+                    if (izq.getValue().tipo == TipoToken.IDENTIFICADOR)
+                    {
                         TablaDeSimbolos.asignar(izq.getValue().lexema, resultadoDerecho);
                     }
                     break;
-
+default:
+break;
             }
         }
-        
-            
-        
-        else if(resultadoIzquierdo instanceof String || resultadoDerecho instanceof String){
-            String valor;
+        else if(resultadoIzquierdo instanceof String && resultadoDerecho instanceof String)
+        {
+            if (n.getValue().tipo == TipoToken.SUMA)
+            {
+                // Ejecutar la concatenación
 
-            if (! (nodo.getValue().tipo == TipoToken.SUMA)){
-                return null;
+                return (String) resultadoIzquierdo.toString() + resultadoDerecho;
+            }
+            else
+            {
+                System.out.println("Error. Solo se permite suma para las cadenas");
+                System.exit(1);
             }
 
-            // Ejecutar la concatenación
-
-            if (!(resultadoDerecho instanceof String)){
-                valor = resultadoDerecho.toString();
-                return resultadoIzquierdo + valor;
-            }
-
-            if (!(resultadoIzquierdo instanceof String)){
-                valor = resultadoIzquierdo.toString();
-                return valor + resultadoDerecho;
-            }
-
-            return (String) resultadoIzquierdo + resultadoDerecho;
         }
+        else if (resultadoIzquierdo instanceof Boolean && resultadoDerecho instanceof Boolean)
+        {
+            switch (n.getValue().tipo)
+            {
+                case AND:
+                    return ((Boolean) resultadoIzquierdo && (Boolean) resultadoDerecho);
+                case OR:
+                    return ((Boolean) resultadoIzquierdo || (Boolean) resultadoDerecho);
+                    default:
+                    break;
+            }
+        }
+        else
+        {
+            // Error por diferencia de tipos
+            System.out.println("Error, no se puede realizar operaciones con variables de diferente tipo");
+            System.exit(1);
+        }
+
         return null;
-    }}
-    
+    }
+}
