+++ AnalizadorSintacticoFinal/AnalizadorSintacticoFinal/TablaDeSimbolos.java
@@ -1,0 +1,30 @@
+package AnalizadorSintacticoFinal;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class TablaDeSimbolos {
+
+    private final static Map<String, Object> values = new HashMap<>();
+
+    static boolean existeIdentificador(String identificador)
+    {
+        return values.containsKey(identificador);
+    }
+
+    static Object obtener(String identificador)
+    {
+        if (values.containsKey(identificador))
+        {
+            return values.get(identificador);
+        }
+        throw new RuntimeException("Variable no definida '" + identificador + "'.");
+    }
+
+    static void asignar(String identificador, Object valor)
+    {
+        values.put(identificador, valor);
+    }
+
+
+}
