+++ AnalizadorSintacticoFinal/AnalizadorSintacticoFinal/TipoToken.java
@@ -1,0 +1,52 @@
+package AnalizadorSintacticoFinal;
+public enum TipoToken {
+    // Palabras reservadas
+    CLASS,
+    FUN,
+    VAR,
+    IF,
+    ELSE,
+    PRINT,
+    RETURN,
+    WHILE,
+    TRUE,
+    FALSE,
+    NULL,
+    THIS,
+    FOR,
+    OR,
+    AND,
+    EXTENDS,
+
+    // Operadores y signos de puntuación
+    ASIGNACION,
+    SUMA,
+    RESTA,
+    MULTIPLICACION,
+    DIVISION,
+    PARENTESIS_IZQ,
+    PARENTESIS_DER,
+    CORCHETE_IZQ,
+    CORCHETE_DER,
+    PUNTO,
+    IGUAL,
+    DIFERENTE,
+    IGUAL_IGUAL,
+    MAYOR_QUE,
+    MAYOR_IGUAL,
+    MENOR_QUE,
+    MENOR_IGUAL,
+    NOT,
+    LLAVE_IZQ,
+    LLAVE_DER,
+
+    // Otros tokens
+    IDENTIFICADOR,
+    NUMERO,
+    CADENA,
+    COMA,
+    PUNTO_Y_COMA,
+    ERROR,
+    SUPER,
+    EOF,
+}
