--- AnalizadorSintacticoFinal/Prueba.txt
@@ -1,5 +1,0 @@
-// Prueba 1:
-var nombre = "Gabriel";
-var apellido = "Rodriguez";
-
-print nombre + " " + apellido;
