--- AnalizadorSintacticoFinal/TablaDeSimbolos.java
@@ -1,25 +1,0 @@
-
-import java.util.HashMap;
-import java.util.Map;
-
-public class TablaDeSimbolos {
-
-    private static final Map<String, Object> values = new HashMap<>();
-
-    static boolean existeIdentificador(String identificador){
-        return values.containsKey(identificador);
-    }
-
-   static Object obtener(String identificador) {
-        if (values.containsKey(identificador)) {
-            return values.get(identificador);
-        }
-        throw new RuntimeException("Variable no definida '" + identificador + "'.");
-    }
-
-   static void asignar(String identificador, Object valor){
-        values.put(identificador, valor);
-    }
-
-
-}
